Limit number of columns in text field
Text field was allowed to take as much space as needed in the scroll
pane. It caused that text field was so long to fit whole text in it. It
is not desired, because it force user to scroll scrollpane horizontally
nearly on each page (if it contains long enough values).